Added installation instructions for FreeNAS.
The installation of Runalyze inside a FreeNAS jail might be a little bit
tricky. Therefore I wrote a short installation instruction. This also
contains a code snippet that deals with perl errors that are printed
when the local is not set correctely. On my box this fixed the fit file
import.